fix server build and requirements.txt
move downloadm3u8 function to video_utils.py

fix imports: typing and pydantic
