+++ jockey/cli.py
@@ -13,58 +13,7 @@ from typing import List
 from langchain_core.runnables.schema import StreamEvent
 from jockey.jockey_graph import FeedbackEntry
 from jockey.thread import session_id, thread
-import json
-from tqdm import tqdm
-import time
-
-
-def download_m3u8_videos(event):
-    """Download and trim M3U8 videos into mp4 files with caching and progress display.
-
-    mp4 files are sorted by start time, then trimmed by end time. they are put into filepath output/{session_id}
-    """
-    # Create session directory if it doesn't exist
-    trimmed_filepath = "output/" + str(session_id)
-    source_filepath = "output/source"
-    os.makedirs(trimmed_filepath, exist_ok=True)
-    os.makedirs(source_filepath, exist_ok=True)
-    downloaded_videos = set()  # Cache of downloaded video URLs for this function call
-
-    for item in tqdm(json.loads(event["data"]["output"]), desc="Processing Videos"):
-        video_url = item.get("video_url")
-        video_id = item.get("video_id")
-
-        if (os.path.exists(os.path.join(source_filepath, f"{video_id}.mp4"))) or (video_id and video_id not in downloaded_videos):
-            start, end = item.get("start", 0), item.get("end")
-            source_file = os.path.join(source_filepath, f"{video_id}.mp4")
-
-            # Download video showing yt-dlp progress
-            subprocess.run(["yt-dlp", "-o", source_file, video_url, "--progress", "--quiet"])
-            downloaded_videos.add(video_id)
-
-        # Trim the video if end time is provided
-        if item.get("end") is not None:
-            start, end = item.get("start", 0), item.get("end")
-            output_file = os.path.join(source_filepath, f"{video_id}.mp4")
-            trimmed_file = os.path.join(trimmed_filepath, f"trimmed-{start}-{end}.mp4")
-            duration = end - start
-
-            # Run ffmpeg silently
-            subprocess.run([
-                "ffmpeg",
-                "-hide_banner",
-                "-loglevel",
-                "error",  # Only show errors
-                "-ss",
-                str(start),
-                "-i",
-                output_file,
-                "-t",
-                str(duration),
-                "-c",
-                "copy",
-                trimmed_file,
-            ])
+from jockey.video_utils import download_m3u8_videos
 
 
 async def run_jockey_terminal():
@@ -91,7 +40,6 @@ async def run_jockey_terminal():
                     # event["chat_history"][-1].pretty_print()
                     events.append(event)
                     if event["event"] == "on_tool_end":
-                        # let's handle the m3u8 video
                         download_m3u8_videos(event)
                     await parse_langchain_events_terminal(event)
