+++ requirements.txt
@@ -3,14 +3,21 @@ aiohttp==3.10.8
 aiosignal==1.3.1
 annotated-types==0.7.0
 anyio==4.6.0
+appnope==0.1.4
 asttokens==2.4.1
 attrs==24.2.0
+backcall==0.2.0
+beautifulsoup4==4.12.3
+bleach==6.2.0
 certifi==2024.8.30
 charset-normalizer==3.3.2
 click==8.1.7
 decorator==5.1.1
+defusedxml==0.7.1
 distro==1.9.0
+docopt==0.6.2
 executing==2.1.0
+fastjsonschema==2.20.0
 ffmpeg-python==0.2.0
 frozenlist==1.4.1
 future==1.0.0
@@ -23,36 +30,50 @@ httpx==0.27.2
 httpx-sse==0.4.0
 idna==3.10
 iniconfig==2.0.0
-ipython==8.29.0
+ipython==8.12.3
 jedi==0.19.1
+Jinja2==3.1.4
 jiter==0.5.0
 jsonpatch==1.33
 jsonpointer==3.0.0
-langchain==0.2.14
-langchain-core==0.2.41
-langchain-openai==0.1.22
-langchain-text-splitters==0.2.4
-langgraph==0.2.39
-langgraph-checkpoint==2.0.1
+jsonschema==4.23.0
+jsonschema-specifications==2024.10.1
+jupyter_client==8.6.3
+jupyter_core==5.7.2
+jupyterlab_pygments==0.3.0
+langchain==0.3.7
+langchain-core==0.3.19
+langchain-openai==0.2.9
+langchain-text-splitters==0.3.2
+langgraph==0.2.52
+langgraph-checkpoint==2.0.5
 langgraph-cli==0.1.52
 langgraph-sdk==0.1.33
 langsmith==0.1.129
 markdown-it-py==3.0.0
+MarkupSafe==3.0.2
 matplotlib-inline==0.1.7
 mdurl==0.1.2
+mistune==3.0.2
 msgpack==1.1.0
 multidict==6.1.0
 mypy-extensions==1.0.0
+nbclient==0.10.0
+nbconvert==7.16.4
+nbformat==5.10.4
 numpy==1.26.4
-openai==1.41.1
+openai==1.54.5
 orjson==3.10.7
 packaging==24.1
+pandocfilters==1.5.1
 parso==0.8.4
 pathspec==0.12.1
 pexpect==4.9.0
+pickleshare==0.7.5
 platformdirs==4.3.6
 pluggy==1.5.0
 prompt_toolkit==3.0.48
+propcache==0.2.0
 ptyprocess==0.7.0
 pure_eval==0.2.3
 pydantic==2.9.2
@@ -60,26 +81,35 @@ pydantic_core==2.23.4
 Pygments==2.18.0
 pytest==8.3.3
 pytest-asyncio==0.24.0
+python-dateutil==2.9.0.post0
 python-dotenv==1.0.1
 PyYAML==6.0.2
+pyzmq==26.2.0
+referencing==0.35.1
 regex==2024.9.11
 requests==2.32.3
-rich==13.7.0
+rich==13.9.4
+rpds-py==0.21.0
 ruff==0.6.9
 setuptools==75.1.0
 six==1.16.0
 sniffio==1.3.1
+soupsieve==2.6
 SQLAlchemy==2.0.35
 stack-data==0.6.3
 tabulate==0.9.0
 tenacity==8.5.0
 tiktoken==0.7.0
+tinycss2==1.4.0
+tornado==6.4.1
 tqdm==4.66.5
 traitlets==5.14.3
 twelvelabs==0.2.8
 typing_extensions==4.12.2
 urllib3==2.2.3
 wcwidth==0.2.13
+webencodings==0.5.1
+yarg==0.1.9
 yarl==1.13.1
 zope.event==5.0
 zope.interface==7.0.3
